Update some notification code
